Versioning & a lot of cleanup

+++ db/manual/remove_all_deleted.go
@@ -1,11 +1,16 @@
 package manual
 
 import (
+	"github.com/rs/zerolog"
 	"gorm.io/gorm"
 	"strings"
 )
 
-func RemoveAllDeleted(db *gorm.DB) error {
+func RemoveAllDeleted(db *gorm.DB, log zerolog.Logger) error {
+	if getCurrentVersion(db) != "" {
+		log.Trace().Msg("Skipping changes, Media-Provider installed after changes are needed")
+		return nil
+	}
 	for _, table := range []string{"modifier_values", "modifiers", "notifications", "pages", "password_resets", "preferences", "subscription_infos", "subscriptions", "tags", "users"} {
 		res := db.Exec("DELETE FROM `" + table + "` WHERE deleted_at IS NOT NULL")
 		if res.Error != nil {
